+++ lib/features/flash_sale/domain/models/product_flash_sale.dart
@@ -42,7 +42,7 @@ class Products {
   int? id;
   int? flashSaleId;
   int? itemId;
-  int? stock;
+  double? stock;
   int? sold;
   int? availableStock;
   String? discountType;
@@ -75,7 +75,7 @@ class Products {
     id = json['id'];
     flashSaleId = json['flash_sale_id'];
     itemId = json['item_id'];
-    stock = json['stock'];
+    stock = json['stock'].toDouble();
     sold = json['sold'];
     availableStock = json['available_stock'];
     discountType = json['discount_type'];
